UPDATED CLEB Files
Project Files & Classes Overview (So Far)

1. Model Layer (com.cleb.model)
Role.java – Enum defining the three user roles (STUDENT, TECHNICIAN, ADMIN)
User.java – Abstract base class for all users (with inheritance and polymorphism)
Student.java, Admin.java, Technician.java – Concrete user types extending User
Lab.java, Seat.java, Equipment.java, Reservation.java – Main business objects (labs, workstations, equipment, and bookings)
StudentModel.java – Helper class for GUI data binding

2. Communication Layer (com.cleb.common)
Request.java – Serialized envelope sent from client to server (contains action and payload + UUID)
Response.java – Serialized reply from server to client
Serializer.java – Utility class for object serialization/deserialization over the network

3. Data Access Layer (com.cleb.dao)
UserDAO.java and ReservationDAO.java – Interfaces (polymorphism)
JdbcUserDAO.java and JdbcReservationDAO.java – Real JDBC implementations that connect to MySQL
HibernateUserDAO.java and HibernateReservationDAO.java – Stubs for the Hibernate version
DatabaseException.java – Custom exception for database errors

4. Server Layer (com.cleb.server)
ServiceFactory.java – Inversion of Control (IoC) class to easily switch between JDBC and Hibernate
ServerMain.java – Starts the TCP server with a thread pool
ClientHandler.java – Handles each connected client (processes requests and sends responses)

5. Client GUI Layer (com.cleb.client)
ClientMain.java – Main application window (JFrame + JDesktopPane MDI)
Session.java – Holds the currently logged-in user
LoginInternalFrame.java – Login form (connects to database)
DataViewerInternalFrame.java – Shows live pending reservations from the database
LabEntryInternalFrame.java, EquipmentEntryInternalFrame.java, UserEntryInternalFrame.java – Data entry forms
ClientListener.java – Background thread that receives real-time updates from the server

6. Configuration & Others
log4j2.xml – Logging configuration (rolling logs)
pom.xml – Maven dependencies

+++ CLEB2/src/main/java/com/cleb/model/Reservation.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class Reservation implements Serializable{
 	private static final long serialVersionUID = 1L;
 	
-	private int reservationId;
+	private long reservationId;
 	private User bookedBy;
 	private String startTime;
 	private String endTime;
@@ -14,7 +14,7 @@ public class Reservation implements Serializable{
 	private String status;
 	
 	
-	public Reservation(int reservationId, User bookedBy, String startTime, String endTime, Seat seat, Equipment equip, String status) {
+	public Reservation(long reservationId, User bookedBy, String startTime, String endTime, Seat seat, Equipment equip, String status) {
 		this.reservationId = reservationId;
         this.bookedBy = bookedBy;
         this.startTime = startTime;
@@ -44,12 +44,12 @@ public class Reservation implements Serializable{
 		this.status = r.status;
 	}
 
-	public int getReservationId() {
+	public long getReservationId() {
 		return reservationId;
 	}
-
-	public void setReservationId(int reservationId) {
-		this.reservationId = reservationId;
+	
+	public void setReservationId(long reservationId) { 
+	    this.reservationId = reservationId; 
 	}
 
 	public User getBookedBy() {
